requirements.txt
pandas
paho.mqtt
influxdb_client